requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annoy==1.17.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
graphviz==0.20.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
openai-secret-manager==0.0.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
